Split pretty_midi.py into separate files
The file pretty_midi.py contained the entire module before; this meant
it was pretty huge.  This commit splits up the module into a few
different files: instrument.py, which contains the Instrument class;
containers.py, which contains the simple "container" classes like Note,
KeySignature, etc.; utilties.py, which contains the non-class utility
functions, and constants.py, which contains the lists used to map MIDI
data constants to readable strings.  pretty_midi.py now simply contains
the PrettyMIDI class.
