+++ pretty_midi/__init__.py
@@ -1,3 +1,7 @@
-from pretty_midi import *
+from .pretty_midi import *
+from .instrument import *
+from .containers import *
+from .utilities import *
+from .constants import *
 
 __version__ = '0.1'
