+++ pretty_midi/pretty_midi.py
@@ -5,19 +5,14 @@ format
 
 import midi
 import numpy as np
-try:
-    import fluidsynth
-    _HAS_FLUIDSYNTH = True
-except ImportError:
-    _HAS_FLUIDSYNTH = False
-import os
 import warnings
-import pkg_resources
-import re
 import collections
 import copy
 
-DEFAULT_SF2 = 'TimGM6mb.sf2'
+from .instrument import Instrument
+from .containers import KeySignature, TimeSignature
+from .containers import Note, PitchBend, ControlChange
+from .utilities import key_name_to_key_number
 
 # The largest we'd ever expect a tick to be
 MAX_TICK = 1e7
@@ -889,1073 +884,3 @@ class PrettyMIDI(object):
         output_pattern.make_ticks_rel()
         # Write it out
         midi.write_midifile(filename, output_pattern)
-
-
-class Instrument(object):
-    """Object to hold event information for a single instrument
-
-    Parameters
-    ----------
-    program : int
-        MIDI program number (instrument index), in [0, 127]
-    is_drum : bool
-        Is the instrument a drum instrument (channel 9)?
-        Default False
-
-    Attributes
-    ----------
-    program : int
-        The program number of this instrument.
-    is_drum : bool
-        Is the instrument a drum instrument (channel 9)?
-    notes : list
-        List of Note objects
-    pitch_bends : list
-        List of of PitchBend objects
-    control_changes : list
-        List of ControlChange objects
-
-    """
-
-    def __init__(self, program, is_drum=False):
-        """Create the Instrument.
-        notes gets initialized to empty list.
-        Fill with `(Instrument).notes.append(event)`
-
-        """
-        self.program = program
-        self.is_drum = is_drum
-        self.notes = []
-        self.pitch_bends = []
-        self.control_changes = []
-
-    def get_onsets(self):
-        """Get all onsets of all notes played by this instrument.
-        May contain duplicates
-
-        Returns
-        -------
-        onsets : np.ndarray
-                List of all note onsets
-
-        """
-        onsets = []
-        # Get the note-on time of each note played by this instrument
-        for note in self.notes:
-            onsets.append(note.start)
-        # Return them sorted (because why not?)
-        return np.sort(onsets)
-
-    def get_piano_roll(self, fs=100, times=None):
-        """Get a piano roll notation of the note events of this instrument.
-
-        Parameters
-        ----------
-        fs : int
-            Sampling frequency of the columns, i.e. each column is spaced apart
-            by 1./fs seconds
-        times : np.ndarray
-            times of the start of each column in the piano roll,
-            Default None which is np.arange(0, get_end_time(), 1./fs)
-
-        Returns
-        -------
-        piano_roll : np.ndarray, shape=(128, times.shape[0])
-            Piano roll matrix of this instrument
-
-        """
-        # If there are no notes, return an empty matrix
-        if self.notes == []:
-            return np.array([[]]*128)
-        # Get the end time of the last event
-        end_time = self.get_end_time()
-        # Extend end time if one was provided
-        if times is not None and times[-1] > end_time:
-            end_time = times[-1]
-        # Allocate a matrix of zeros - we will add in as we go
-        piano_roll = np.zeros((128, int(fs*end_time)), dtype=np.int16)
-        # Drum tracks don't have pitch, so return a matrix of zeros
-        if self.is_drum:
-            if times is None:
-                return piano_roll
-            else:
-                return np.zeros((128, times.shape[0]), dtype=np.int16)
-        # Add up piano roll matrix, note-by-note
-        for note in self.notes:
-            # Should interpolate
-            piano_roll[note.pitch,
-                       int(note.start*fs):int(note.end*fs)] += note.velocity
-
-        # Process pitch changes
-        # Need to sort the pitch bend list for the following to work
-        ordered_bends = sorted(self.pitch_bends, key=lambda bend: bend.time)
-        # Add in a bend of 0 at the end of time
-        end_bend = PitchBend(0, end_time)
-        for start_bend, end_bend in zip(ordered_bends,
-                                        ordered_bends[1:] + [end_bend]):
-            # Piano roll is already generated with everything bend = 0
-            if np.abs(start_bend.pitch) < 1:
-                continue
-            # Get integer and decimal part of bend amount
-            start_pitch = pitch_bend_to_semitones(start_bend.pitch)
-            bend_int = int(np.sign(start_pitch)*np.floor(np.abs(start_pitch)))
-            bend_decimal = np.abs(start_pitch - bend_int)
-            # Column indices effected by the bend
-            bend_range = np.r_[int(start_bend.time*fs):int(end_bend.time*fs)]
-            # Construct the bent part of the piano roll
-            bent_roll = np.zeros(piano_roll[:, bend_range].shape)
-            # Easiest to process differently depending on bend sign
-            if start_bend.pitch >= 0:
-                # First, pitch shift by the int amount
-                if bend_int is not 0:
-                    bent_roll[bend_int:] = piano_roll[:-bend_int, bend_range]
-                else:
-                    bent_roll = piano_roll[:, bend_range]
-                # Now, linear interpolate by the decimal place
-                bent_roll[1:] = ((1 - bend_decimal)*bent_roll[1:]
-                                 + bend_decimal*bent_roll[:-1])
-            else:
-                # Same procedure as for positive bends
-                if bend_int is not 0:
-                    bent_roll[:bend_int] = piano_roll[-bend_int:, bend_range]
-                else:
-                    bent_roll = piano_roll[:, bend_range]
-                bent_roll[:-1] = ((1 - bend_decimal)*bent_roll[:-1]
-                                  + bend_decimal*bent_roll[1:])
-            # Store bent portion back in piano roll
-            piano_roll[:, bend_range] = bent_roll
-
-        if times is None:
-            return piano_roll
-        piano_roll_integrated = np.zeros((128, times.shape[0]), dtype=np.int16)
-        # Convert to column indices
-        times = np.array(times*fs, dtype=np.int)
-        for n, (start, end) in enumerate(zip(times[:-1], times[1:])):
-            # Each column is the mean of the columns in piano_roll
-            piano_roll_integrated[:, n] = np.mean(piano_roll[:, start:end],
-                                                  axis=1)
-        return piano_roll_integrated
-
-    def get_chroma(self, fs=100, times=None):
-        """Get a chroma matrix for the note events in this instrument.
-
-        Parameters
-        ----------
-        fs : int
-            Sampling frequency of the columns, i.e. each column is spaced
-            apart by 1./fs seconds
-        times : np.ndarray
-            times of the start of each column in the chroma matrix,
-            Default None which is np.arange(0, get_end_time(), 1./fs)
-
-        Returns
-        -------
-        chroma : np.ndarray, shape=(12,times.shape[0])
-            Chromagram matrix of this instrument
-
-        """
-        # First, get the piano roll
-        piano_roll = self.get_piano_roll(fs=fs, times=times)
-        # Fold into one octave
-        chroma_matrix = np.zeros((12, piano_roll.shape[1]))
-        for note in range(12):
-            chroma_matrix[note, :] = np.sum(piano_roll[note::12], axis=0)
-        return chroma_matrix
-
-    def get_end_time(self):
-        """Returns the time of the end of the events in this instrument
-
-        Returns
-        -------
-        end_time : float
-            Time, in seconds, of the end of the last event
-
-        """
-        # Cycle through all note ends and all pitch bends and find the largest
-        events = ([n.end for n in self.notes] +
-                  [b.time for b in self.pitch_bends])
-        # If there are no events, just return 0
-        if len(events) == 0:
-            return 0.
-        else:
-            return max(events)
-
-    def remove_invalid_notes(self):
-        """Removes any notes which have an end time <= start time.
-
-        """
-        # Crete a list of all invalid notes
-        notes_to_delete = []
-        for note in self.notes:
-            if note.end <= note.start:
-                notes_to_delete.append(note)
-        # Remove the notes found
-        for note in notes_to_delete:
-            self.notes.remove(note)
-
-    def synthesize(self, fs=44100, wave=np.sin):
-        """Synthesize the instrument's notes using some waveshape.
-        For drum instruments, returns zeros.
-
-        Parameters
-        ----------
-        fs : int
-            Sampling rate of the synthesized audio signal, default 44100
-        wave : function
-            Function which returns a periodic waveform,
-            e.g. np.sin, scipy.signal.square, etc.  Default np.sin
-
-        Returns
-        -------
-        synthesized : np.ndarray
-            Waveform of the instrument's notes, synthesized at fs
-
-        """
-        # Pre-allocate output waveform
-        synthesized = np.zeros(int(fs*(self.get_end_time() + 1)))
-
-        # If we're a percussion channel, just return the zeros
-        if self.is_drum:
-            return synthesized
-        # If the above if statement failed, we need to revert back to default
-        if not hasattr(wave, '__call__'):
-            raise ValueError('wave should be a callable Python function')
-        # This is a simple way to make the end of the notes fade-out without
-        # clicks
-        fade_out = np.linspace(1, 0, .1*fs)
-        # Create a frequency multiplier array for pitch bend
-        bend_multiplier = np.ones(synthesized.shape)
-        # Need to sort the pitch bend list for the loop below to work
-        ordered_bends = sorted(self.pitch_bends, key=lambda bend: bend.time)
-        # Add in a bend of 0 at the end of time
-        end_bend = PitchBend(0, self.get_end_time())
-        for start_bend, end_bend in zip(ordered_bends,
-                                        ordered_bends[1:] + [end_bend]):
-            # Bend start and end time in samples
-            start = int(start_bend.time*fs)
-            end = int(end_bend.time*fs)
-            # The multiplier will be (twelfth root of 2)^(bend semitones)
-            bend_semitones = pitch_bend_to_semitones(start_bend.pitch)
-            bend_amount = (2**(1/12.))**bend_semitones
-            # Sample indices effected by the bend
-            bend_multiplier[start:end] = bend_amount
-        # Add in waveform for each note
-        for note in self.notes:
-            # Indices in samples of this note
-            start = int(fs*note.start)
-            end = int(fs*note.end)
-            # Get frequency of note from MIDI note number
-            frequency = note_number_to_hz(note.pitch)
-            # When a pitch bend gets applied, there will be a sample
-            # discontinuity. So, we also need an array of offsets which get
-            # applied to compensate.
-            offsets = np.zeros(end - start)
-            for bend in ordered_bends:
-                bend_sample = int(bend.time*fs)
-                # Does this pitch bend fall within this note?
-                if bend_sample > start and bend_sample < end:
-                    # Compute the average bend so far
-                    bend_so_far = bend_multiplier[start:bend_sample].mean()
-                    bend_amount = bend_multiplier[bend_sample]
-                    # Compute the offset correction
-                    offset = (bend_so_far - bend_amount)*(bend_sample - start)
-                    # Store this offset for samples effected
-                    offsets[bend_sample - start:] = offset
-            # Compute the angular frequencies, bent, over this interval
-            frequencies = 2*np.pi*frequency*(bend_multiplier[start:end])/fs
-            # Synthesize using wave function at this frequency
-            note_waveform = wave(frequencies*np.arange(end - start) +
-                                 2*np.pi*frequency*offsets/fs)
-            # Apply an exponential envelope
-            envelope = np.exp(-np.arange(end - start)/(1.0*fs))
-            # Make the end of the envelope be a fadeout
-            if envelope.shape[0] > fade_out.shape[0]:
-                envelope[-fade_out.shape[0]:] *= fade_out
-            else:
-                envelope *= np.linspace(1, 0, envelope.shape[0])
-            # Multiply by velocity (don't think it's linearly scaled but
-            # whatever)
-            envelope *= note.velocity
-            # Add in envelope'd waveform to the synthesized signal
-            synthesized[start:end] += envelope*note_waveform
-
-        return synthesized
-
-    def fluidsynth(self, fs=44100, sf2_path=None):
-        """Synthesize using fluidsynth.
-
-        Parameters
-        ----------
-        fs : int
-            Sampling rate to synthesize
-        sf2_path : str
-            Path to a .sf2 file.
-            Default None, which uses the TimGM6mb.sf2 file included with
-            pretty_midi.
-
-        Returns
-        -------
-        synthesized : np.ndarray
-            Waveform of the MIDI data, synthesized at fs
-
-        """
-        # If sf2_path is None, use the included TimGM6mb.sf2 path
-        if sf2_path is None:
-            sf2_path = pkg_resources.resource_filename(__name__, DEFAULT_SF2)
-
-        if not _HAS_FLUIDSYNTH:
-            raise ImportError("fluidsynth() was called but pyfluidsynth "
-                              "is not installed.")
-
-        if not os.path.exists(sf2_path):
-            raise ValueError("No soundfont file found at the supplied path "
-                             "{}".format(sf2_path))
-
-        # Create fluidsynth instance
-        fl = fluidsynth.Synth(samplerate=fs)
-        # Load in the soundfont
-        sfid = fl.sfload(sf2_path)
-        # If this is a drum instrument, use channel 9 and bank 128
-        if self.is_drum:
-            channel = 9
-            # Try to use the supplied program number
-            res = fl.program_select(channel, sfid, 128, self.program)
-            # If the result is -1, there's no preset with this program number
-            if res == -1:
-                # So use preset 0
-                fl.program_select(channel, sfid, 128, 0)
-        # Otherwise just use channel 0
-        else:
-            channel = 0
-            fl.program_select(channel, sfid, 0, self.program)
-        # Collect all notes in one list
-        event_list = []
-        for note in self.notes:
-            event_list += [[note.start, 'note on', note.pitch, note.velocity]]
-            event_list += [[note.end, 'note off', note.pitch]]
-        for bend in self.pitch_bends:
-            event_list += [[bend.time, 'pitch bend', bend.pitch]]
-        for control_change in self.control_changes:
-            event_list += [[control_change.time, 'control change',
-                            control_change.number, control_change.value]]
-        # Sort the event list by time, and secondarily by whether the event
-        # is a note off
-        event_list.sort(key=lambda x: (x[0], x[1] != 'note off'))
-        # Add some silence at the beginning according to the time of the first
-        # event
-        current_time = event_list[0][0]
-        # Convert absolute seconds to relative samples
-        next_event_times = [e[0] for e in event_list[1:]]
-        for event, end in zip(event_list[:-1], next_event_times):
-            event[0] = end - event[0]
-        # Include 1 second of silence at the end
-        event_list[-1][0] = 1.
-        # Pre-allocate output array
-        total_time = current_time + np.sum([e[0] for e in event_list])
-        synthesized = np.zeros(int(np.ceil(fs*total_time)))
-        # Iterate over all events
-        for event in event_list:
-            # Process events based on type
-            if event[1] == 'note on':
-                fl.noteon(channel, event[2], event[3])
-            elif event[1] == 'note off':
-                fl.noteoff(channel, event[2])
-            elif event[1] == 'pitch bend':
-                fl.pitch_bend(channel, event[2])
-            elif event[1] == 'control change':
-                fl.cc(channel, event[2], event[3])
-            # Add in these samples
-            current_sample = int(fs*current_time)
-            end = int(fs*(current_time + event[0]))
-            samples = fl.get_samples(end - current_sample)[::2]
-            synthesized[current_sample:end] += samples
-            # Increment the current sample
-            current_time += event[0]
-        # Close fluidsynth
-        fl.delete()
-
-        return synthesized
-
-    def __repr__(self):
-        return 'Instrument(program={}, is_drum={})'.format(
-            self.program, self.is_drum, len(self.notes))
-
-
-class Note(object):
-    """A note event.
-
-    Parameters
-    ----------
-    velocity : int
-        Note velocity
-    pitch : int
-        Note pitch, as a MIDI note number
-    start : float
-        Note on time, absolute, in seconds
-    end : float
-        Note off time, absolute, in seconds
-
-    Notes
-    -----
-    This class is simply a container for its parameters.
-
-    """
-
-    def __init__(self, velocity, pitch, start, end):
-        """Create a note object.
-
-        """
-        self.velocity = velocity
-        self.pitch = pitch
-        self.start = start
-        self.end = end
-
-    def __repr__(self):
-        return 'Note(start={:f}, end={:f}, pitch={}, velocity={})'.format(
-            self.start, self.end, self.pitch, self.velocity)
-
-
-class PitchBend(object):
-    """A pitch bend event.
-
-    Parameters
-    ----------
-    pitch : int
-        MIDI pitch bend amount, in the range [-8192, 8191]
-    time : float
-        Time where the pitch bend occurs
-
-    Notes
-    -----
-    This class is simply a container for its parameters.
-
-    """
-
-    def __init__(self, pitch, time):
-        """Create pitch bend object.
-
-        """
-        self.pitch = pitch
-        self.time = time
-
-    def __repr__(self):
-        return 'PitchBend(pitch={:d}, time={:f})'.format(self.pitch, self.time)
-
-
-class ControlChange(object):
-    """A control change event.
-
-    Parameters
-    ----------
-    number : int
-        The control change number, in [0, 127]
-    value : int
-        The value of the control change, in [0, 127]
-    time : float
-        Time where the pitch bend occurs
-
-    Notes
-    -----
-    This class is simply a container for its parameters.
-
-    """
-
-    def __init__(self, number, value, time):
-        """Create control change object.
-
-        """
-        self.number = number
-        self.value = value
-        self.time = time
-
-    def __repr__(self):
-        return ('ControlChange(number={:d}, value={:d}, '
-                'time={:f})'.format(self.number, self.value, self.time))
-
-
-class TimeSignature(object):
-    """Container for a Time Signature event, which contains the time signature
-    numerator, denominator and the event time in seconds
-
-    Attributes
-    ----------
-        numerator : int
-            Numerator of time signature
-        denominator : int
-            Denominator of time signature
-        time : float
-            Time of event in seconds
-
-    Examples
-    --------
-    Instantiate a TimeSignature object with 6/8 time signature at 3.14 seconds
-    >>> ts = TimeSignature(6, 8, 3.14)
-    >>> print ts
-    6/8 at 3.14 seconds
-
-    """
-
-    def __init__(self, numerator, denominator, time):
-        if not (isinstance(numerator, int) and numerator > 0):
-            raise ValueError(
-                '{} is not a valid `numerator` type or value'.format(
-                    numerator))
-        if not (isinstance(denominator, int) and denominator > 0):
-            raise ValueError(
-                '{} is not a valid `denominator` type or value'.format(
-                    denominator))
-        if not (isinstance(time, (int, float)) and time >= 0):
-            raise ValueError(
-                '{} is not a valid `time` type or value'.format(time))
-
-        self.numerator = numerator
-        self.denominator = denominator
-        self.time = time
-
-    def __repr__(self):
-        return "TimeSignature(numerator={}, denominator={}, time={})".format(
-            self.numerator, self.denominator, self.time)
-
-    def __str__(self):
-        return '{}/{} at {:.2f} seconds'.format(
-            self.numerator, self.denominator, self.time)
-
-
-class KeySignature(object):
-    """Contains the key signature and the event time in seconds
-
-    Only supports major and minor keys.
-
-    Attributes
-    ----------
-    key_number : int
-        key number accordingly to [0,11] Major, [12,23] minor
-        For example, 0 is C Major, 12 is C minor
-    time : float
-        time of event in seconds
-
-    Examples
-    --------
-    Instantiate a C# minor KeySignature object at 3.14 seconds.
-
-    >>> ks = KeySignature(13, 3.14)
-    >>> print ks
-    C# minor at 3.14 seconds
-    """
-
-    def __init__(self, key_number, time):
-        if not all((isinstance(key_number, int),
-                    key_number >= 0,
-                    key_number < 24)):
-            raise ValueError(
-                '{} is not a valid `key_number` type or value'.format(
-                    key_number))
-        if not (isinstance(time, (int, float)) and time >= 0):
-            raise ValueError(
-                '{} is not a valid `time` type or value'.format(time))
-
-        self.key_number = key_number
-        self.time = time
-
-    def __repr__(self):
-        return "KeySignature(key_number={}, time={})".format(
-            self.key_number, self.time)
-
-    def __str__(self):
-        return '{} at {:.2f} seconds'.format(
-            key_number_to_key_name(self.key_number), self.time)
-
-
-def key_number_to_key_name(key_number):
-    """Convert a key number to a key string
-
-    Parameters
-    ----------
-    key_number : int
-        Uses pitch classes to represent major and minor keys.
-        For minor keys, adds a 12 offset.
-        For example, C major is 0 and C minor is 12.
-
-    Returns
-    -------
-    str
-        'Root mode', e.g. Gb minor.
-        Gives preference for keys with flats, with the
-        exception of F#, G# and C# minor.
-    """
-
-    if not isinstance(key_number, int):
-        raise ValueError('`key_number` is not int!')
-    if not ((key_number >= 0) and (key_number < 24)):
-        raise ValueError('`key_number` is larger than 24')
-
-    # preference to keys with flats
-    keys = ['C', 'Db', 'D', 'Eb', 'E', 'F', 'Gb',
-            'G', 'Ab', 'A', 'Bb', 'B']
-
-    # circle around 12 pitch classes
-    key_idx = key_number % 12
-    mode = key_number / 12
-
-    # check if mode is major or minor
-    if mode == 0:
-        return keys[key_idx] + ' Major'
-    elif mode == 1:
-        # preference to C#, F# and G# minor
-        if key_idx in [1, 6, 8]:
-            return keys[key_idx-1] + '# minor'
-        else:
-            return keys[key_idx] + ' minor'
-
-
-def key_name_to_key_number(key_string):
-    """Convert a correctly formated string key to key number
-
-    Parameters
-    ----------
-    key_string : str
-        Format is ``'key mode'``, where:
-            `key` is notaded using ABCDEFG with # or b;
-            `mode` is notated using 'major' or 'minor'.
-        Letter case is irrelevant for mode.
-    """
-
-    if not isinstance(key_string, str):
-        raise ValueError('key_string is not String')
-    if not key_string[1] in ['#', 'b', ' ']:
-        raise ValueError(
-            '2nd character {} is not #, b nor blank_space'.format(
-                key_string[1]))
-
-    # split key and mode, ignore case
-    key_str, mode_str = key_string.split()
-    key_str = key_str.upper()
-    mode_str = mode_str.lower()
-
-    # instantiate default pitch classes and supported modes
-    note_names_pc = {'C': 0, 'D': 2, 'E': 4,
-                     'F': 5, 'G': 7, 'A': 9, 'B': 11}
-    modes = ['major', 'minor']
-
-    # check that both key and mode are valid
-    if not key_str[0] in note_names_pc:
-        raise ValueError('Key {} is not recognized'.format(key_str[0]))
-    if mode_str not in modes:
-        raise ValueError('Mode {} is not recognized'.format(mode_str))
-
-    # lookup dictionary
-    key_number = note_names_pc[key_str[0]]
-
-    # if len is 2, has a sharp or flat
-    if len(key_str) == 2:
-        if key_str[1] == '#':
-            key_number += 1
-        else:
-            key_number -= 1
-
-    # circle around 12 pitch classes
-    key_number = key_number % 12
-
-    # offset if mode is minor
-    if mode_str == 'minor':
-        key_number += 12
-
-    return key_number
-
-
-def note_number_to_hz(note_number):
-    """Convert a (fractional) MIDI note number to its frequency in Hz.
-
-    Parameters
-    ----------
-    note_number : float
-        MIDI note number, can be fractional
-
-    Returns
-    -------
-    note_frequency : float
-        Frequency of the note in Hz
-
-    """
-    # MIDI note numbers are defined as the number of semitones relative to C0
-    # in a 440 Hz tuning
-    return 440.0*(2.0**((note_number - 69)/12.0))
-
-
-def hz_to_note_number(frequency):
-    """Convert a frequency in Hz to a (fractional) frequency
-
-    Parameters
-    ----------
-    frequency : float
-        Frequency of the note in Hz
-
-    Returns
-    -------
-    note_number : float
-        MIDI note number, can be fractional
-
-    """
-    # MIDI note numbers are defined as the number of semitones relative to C0
-    # in a 440 Hz tuning
-    return 12*(np.log2(frequency) - np.log2(440.0)) + 69
-
-
-def note_name_to_number(note_name):
-    """Converts a note name in the format (note)(accidental)(octave number)
-    to MIDI note number.
-
-    (note) is required, and is case-insensitive.
-
-    (accidental) should be '' for natural, '#' for sharp and '!' or 'b' for
-    flat.
-
-    If (octave) is '', octave 0 is assumed.
-
-    Parameters
-    ----------
-    note_name : str
-        A note name, as described above
-
-    Returns
-    -------
-    note_number : int
-        MIDI note number corresponding to the provided note name.
-
-    Notes
-    -----
-        Thanks to Brian McFee.
-
-    """
-
-    # Map note name to the semitone
-    pitch_map = {'C': 0, 'D': 2, 'E': 4, 'F': 5, 'G': 7, 'A': 9, 'B': 11}
-    # Relative change in semitone denoted by each accidental
-    acc_map = {'#': 1, '': 0, 'b': -1, '!': -1}
-
-    # Reg exp will raise an error when the note name is not valid
-    try:
-        # Extract pitch, octave, and accidental from the supplied note name
-        match = re.match(r'^(?P<n>[A-Ga-g])(?P<off>[#b!]?)(?P<oct>[+-]?\d+)$',
-                         note_name)
-
-        pitch = match.group('n').upper()
-        offset = acc_map[match.group('off')]
-        octave = int(match.group('oct'))
-    except:
-        raise ValueError('Improper note format: {}'.format(note_name))
-
-    # Convert from the extrated ints to a full note number
-    return 12*octave + pitch_map[pitch] + offset
-
-
-def note_number_to_name(note_number):
-    """Convert a MIDI note number to its name, in the format
-    (note)(accidental)(octave number) (e.g. 'C#4')
-
-    Parameters
-    ----------
-    note_number : int
-        MIDI note number.  If not an int, it will be rounded.
-
-    Returns
-    -------
-    note_name : str
-        Name of the supplied MIDI note number.
-
-    Notes
-    -----
-        Thanks to Brian McFee.
-
-    """
-
-    # Note names within one octave
-    semis = ['C', 'C#', 'D', 'D#', 'E', 'F', 'F#', 'G', 'G#', 'A', 'A#', 'B']
-
-    # Ensure the note is an int
-    note_number = int(np.round(note_number))
-
-    # Get the semitone and the octave, and concatenate to create the name
-    return semis[note_number % 12] + str(note_number/12)
-
-# List which maps MIDI note number - 35 to drum name
-# from http://www.midi.org/techspecs/gm1sound.php
-__DRUM_MAP = ['Acoustic Bass Drum', 'Bass Drum 1', 'Side Stick',
-              'Acoustic Snare', 'Hand Clap', 'Electric Snare',
-              'Low Floor Tom', 'Closed Hi Hat', 'High Floor Tom',
-              'Pedal Hi Hat', 'Low Tom', 'Open Hi Hat',
-              'Low-Mid Tom', 'Hi-Mid Tom', 'Crash Cymbal 1',
-              'High Tom', 'Ride Cymbal 1', 'Chinese Cymbal',
-              'Ride Bell', 'Tambourine', 'Splash Cymbal',
-              'Cowbell', 'Crash Cymbal 2', 'Vibraslap',
-              'Ride Cymbal 2', 'Hi Bongo', 'Low Bongo',
-              'Mute Hi Conga', 'Open Hi Conga', 'Low Conga',
-              'High Timbale', 'Low Timbale', 'High Agogo',
-              'Low Agogo', 'Cabasa', 'Maracas',
-              'Short Whistle', 'Long Whistle', 'Short Guiro',
-              'Long Guiro', 'Claves', 'Hi Wood Block',
-              'Low Wood Block', 'Mute Cuica', 'Open Cuica',
-              'Mute Triangle', 'Open Triangle']
-
-
-def note_number_to_drum_name(note_number):
-    """Converts a MIDI note number in a percussion instrument to the
-    corresponding drum name, according to the General MIDI standard.
-
-    Any MIDI note number outside of the valid range (note 35-81, zero-indexed)
-    will result in an empty string.
-
-    Parameters
-    ----------
-    note_number : int
-        MIDI note number.  If not an int, it will be rounded.
-
-    Returns
-    -------
-    drum_name : str
-        Name of the drum for this note for a percussion instrument.
-
-    Notes
-    -----
-        See http://www.midi.org/techspecs/gm1sound.php
-
-    """
-
-    # Ensure note is an int
-    note_number = int(np.round(note_number))
-    # General MIDI only defines drum names for notes 35-81
-    if note_number < 35 or note_number > 81:
-        return ''
-    else:
-        # Our __DRUM_MAP starts from index 0; drum names start from 35
-        return __DRUM_MAP[note_number - 35]
-
-
-def __normalize_str(name):
-    """Removes all non-alphanumeric characters from a string and converts
-    it to lowercase.
-
-    """
-    return ''.join(ch for ch in name if ch.isalnum()).lower()
-
-
-def drum_name_to_note_number(drum_name):
-    """Converts a drum name to the corresponding MIDI note number for a
-    percussion instrument.  Conversion is case, whitespace, and
-    non-alphanumeric character insensitive.
-
-    Parameters
-    ----------
-    drum_name : str
-        Name of a drum which exists in the general MIDI standard.
-        If the drum is not found, a ValueError is raised.
-
-    Returns
-    -------
-    note_number : int
-        The MIDI note number corresponding to this drum.
-
-    Notes
-    -----
-        See http://www.midi.org/techspecs/gm1sound.php
-
-    """
-
-    normalized_drum_name = __normalize_str(drum_name)
-    # Create a list of the entries __DRUM_MAP, normalized, to search over
-    normalized_drum_names = [__normalize_str(name) for name in __DRUM_MAP]
-
-    # If the normalized drum name is not found, complain
-    try:
-        note_index = normalized_drum_names.index(normalized_drum_name)
-    except:
-        raise ValueError('{} is not a valid General MIDI drum '
-                         'name.'.format(drum_name))
-
-    # If an index was found, it will be 0-based; add 35 to get the note number
-    return note_index + 35
-
-
-__INSTRUMENT_MAP = ['Acoustic Grand Piano', 'Bright Acoustic Piano',
-                    'Electric Grand Piano', 'Honky-tonk Piano',
-                    'Electric Piano 1', 'Electric Piano 2', 'Harpsichord',
-                    'Clavinet', 'Celesta', 'Glockenspiel', 'Music Box',
-                    'Vibraphone', 'Marimba', 'Xylophone', 'Tubular Bells',
-                    'Dulcimer', 'Drawbar Organ', 'Percussive Organ',
-                    'Rock Organ', 'Church Organ', 'Reed Organ', 'Accordion',
-                    'Harmonica', 'Tango Accordion', 'Acoustic Guitar (nylon)',
-                    'Acoustic Guitar (steel)', 'Electric Guitar (jazz)',
-                    'Electric Guitar (clean)', 'Electric Guitar (muted)',
-                    'Overdriven Guitar', 'Distortion Guitar',
-                    'Guitar Harmonics', 'Acoustic Bass',
-                    'Electric Bass (finger)', 'Electric Bass (pick)',
-                    'Fretless Bass', 'Slap Bass 1', 'Slap Bass 2',
-                    'Synth Bass 1', 'Synth Bass 2', 'Violin', 'Viola', 'Cello',
-                    'Contrabass', 'Tremolo Strings', 'Pizzicato Strings',
-                    'Orchestral Harp', 'Timpani', 'String Ensemble 1',
-                    'String Ensemble 2', 'Synth Strings 1', 'Synth Strings 2',
-                    'Choir Aahs', 'Voice Oohs', 'Synth Choir', 'Orchestra Hit',
-                    'Trumpet', 'Trombone', 'Tuba', 'Muted Trumpet',
-                    'French Horn', 'Brass Section', 'Synth Brass 1',
-                    'Synth Brass 2', 'Soprano Sax', 'Alto Sax', 'Tenor Sax',
-                    'Baritone Sax', 'Oboe', 'English Horn', 'Bassoon',
-                    'Clarinet', 'Piccolo', 'Flute', 'Recorder', 'Pan Flute',
-                    'Blown bottle', 'Shakuhachi', 'Whistle', 'Ocarina',
-                    'Lead 1 (square)', 'Lead 2 (sawtooth)',
-                    'Lead 3 (calliope)', 'Lead 4 chiff', 'Lead 5 (charang)',
-                    'Lead 6 (voice)', 'Lead 7 (fifths)',
-                    'Lead 8 (bass + lead)', 'Pad 1 (new age)', 'Pad 2 (warm)',
-                    'Pad 3 (polysynth)', 'Pad 4 (choir)', 'Pad 5 (bowed)',
-                    'Pad 6 (metallic)', 'Pad 7 (halo)', 'Pad 8 (sweep)',
-                    'FX 1 (rain)', 'FX 2 (soundtrack)', 'FX 3 (crystal)',
-                    'FX 4 (atmosphere)', 'FX 5 (brightness)', 'FX 6 (goblins)',
-                    'FX 7 (echoes)', 'FX 8 (sci-fi)', 'Sitar', 'Banjo',
-                    'Shamisen', 'Koto', 'Kalimba', 'Bagpipe', 'Fiddle',
-                    'Shanai', 'Tinkle Bell', 'Agogo', 'Steel Drums',
-                    'Woodblock', 'Taiko Drum', 'Melodic Tom', 'Synth Drum',
-                    'Reverse Cymbal', 'Guitar Fret Noise', 'Breath Noise',
-                    'Seashore', 'Bird Tweet', 'Telephone Ring', 'Helicopter',
-                    'Applause', 'Gunshot']
-
-
-def program_to_instrument_name(program_number):
-    """Converts a MIDI program number to the corresponding General MIDI
-    instrument name.
-
-    Parameters
-    ----------
-    program_number : int
-        MIDI program number, between 0 and 127
-
-    Returns
-    -------
-    instrument_name : str
-        Name of the instrument corresponding to this program number.
-
-    Notes
-    -----
-        See http://www.midi.org/techspecs/gm1sound.php
-
-    """
-
-    # Check that the supplied program is in the valid range
-    if program_number < 0 or program_number > 127:
-        raise ValueError('Invalid program number {}, should be between 0 and'
-                         ' 127'.format(program_number))
-    # Just grab the name from the instrument mapping list
-    return __INSTRUMENT_MAP[program_number]
-
-
-def instrument_name_to_program(instrument_name):
-    """Converts an instrument name to the corresponding General MIDI program
-    number.  Conversion is case, whitespace, and non-alphanumeric character
-    insensitive.
-
-    Parameters
-    ----------
-    instrument_name : str
-        Name of an instrument which exists in the general MIDI standard.
-        If the instrument is not found, a ValueError is raised.
-
-    Returns
-    -------
-    program_number : int
-        The MIDI program number corresponding to this instrument.
-
-    Notes
-    -----
-        See http://www.midi.org/techspecs/gm1sound.php
-
-    """
-
-    normalized_inst_name = __normalize_str(instrument_name)
-    # Create a list of the entries __INSTRUMENT_MAP, normalized, to search over
-    normalized_inst_names = [__normalize_str(name) for name in
-                             __INSTRUMENT_MAP]
-
-    # If the normalized drum name is not found, complain
-    try:
-        program_number = normalized_inst_names.index(normalized_inst_name)
-    except:
-        raise ValueError('{} is not a valid General MIDI instrument '
-                         'name.'.format(instrument_name))
-
-    # Return the index (program number) if a match was found
-    return program_number
-
-
-__INSTRUMENT_CLASSES = ['Piano', 'Chromatic Percussion', 'Organ', 'Guitar',
-                        'Bass', 'Strings', 'Ensemble', 'Brass', 'Reed', 'Pipe',
-                        'Synth Lead', 'Synth Pad', 'Synth Effects', 'Ethnic',
-                        'Percussive',
-                        'Sound Effects']
-
-
-def program_to_instrument_class(program_number):
-    """Converts a MIDI program number to the corresponding General MIDI
-    instrument class.
-
-    Parameters
-    ----------
-    program_number : int
-        MIDI program number, between 0 and 127
-
-    Returns
-    -------
-    instrument_class : str
-        Name of the instrument class corresponding to this program number.
-
-    Notes
-    -----
-        See http://www.midi.org/techspecs/gm1sound.php
-
-    """
-
-    # Check that the supplied program is in the valid range
-    if program_number < 0 or program_number > 127:
-        raise ValueError('Invalid program number {}, should be between 0 and'
-                         ' 127'.format(program_number))
-    # Just grab the name from the instrument mapping list
-    return __INSTRUMENT_CLASSES[int(program_number)/8]
-
-
-def pitch_bend_to_semitones(pitch_bend, semitone_range=2.):
-    """Convert a MIDI pitch bend value (in the range -8192, 8191) to the bend
-    amount in semitones.
-
-    Parameters
-    ----------
-    pitch_bend : int
-        MIDI pitch bend amount, in [-8192, 8191]
-    semitone_range : float
-        Convert to +/- this semitone range.  Default is 2., which is the
-        General MIDI standard +/-2 semitone range.
-
-    Returns
-    -------
-    semitones : float
-        Number of semitones corresponding to this pitch bend amount
-
-    """
-
-    return semitone_range*pitch_bend/8192.0
-
-
-def semitones_to_pitch_bend(semitones, semitone_range=2.):
-    """Convert a semitone value to the corresponding MIDI pitch bend int
-
-    Parameters
-    ----------
-    semitones : float
-        Number of semitones for the pitch bend
-    semitone_range : float
-        Convert to +/- this semitone range.  Default is 2., which is the
-        General MIDI standard +/-2 semitone range.
-
-    Returns
-    -------
-    pitch_bend : int
-        MIDI pitch bend amount, in [-8192, 8191]
-
-    """
-    return int(8192*(semitones/semitone_range))
